iwlwifi: disable greenfield transmissions as a workaround
Patchwork-id: 51176
O-Subject: [RHEL6.4 PATCH 36/38] iwlwifi: disable greenfield transmissions as a workaround
Bugzilla: 841578
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>

From: Johannes Berg <johannes.berg@intel.com>

BZ#841578

commit 50e2a30cf6fcaeb2d27360ba614dd169a10041c5 upstream.

There's a bug that causes the rate scaling to get stuck
when it has to use single-stream rates with a peer that
can do GF and SGI; the two are incompatible so we can't
use them together, but that causes the algorithm to not
work at all, it always rejects updates.

Disable greenfield for now to prevent that problem.

Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Tested-by: Cesar Eduardo Barros <cesarb@cesarb.net>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Signed-off-by: John W. Linville <linville@tuxdriver.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: Jarod Wilson <jarod@redhat.com>