raid1: don't abort a resync on the first badblock
Patchwork-id: 51225
O-Subject: [RHEL6.4 PATCH 1/1] md/raid1: don't abort a resync on the first badblock.
Bugzilla: 858219
RH-Acked-by: Doug Ledford <dledford@redhat.com>
RH-Acked-by: Jonathan E Brassow <jbrassow@redhat.com>

From: NeilBrown <neilb@suse.de>

RHEL6.4's drivers/md stack was resync'ed to upstream 3.5 in
BZ#817970. This is a stable-fix that went into linux-3.5 stable post
the resync.

Bugzilla:
  https://bugzilla.redhat.com/show_bug.cgi?id=858219
Brew-build:
  http://brewweb.devel.redhat.com/brew/taskinfo?taskID=4880234

If a resync of a RAID1 array with 2 devices finds a known bad block
one device it will neither read from, or write to, that device for
this block offset.
So there will be one read_target (The other device) and zero write
targets.
This condition causes md/raid1 to abort the resync assuming that it
has finished - without known bad blocks this would be true.

When there are no write targets because of the presence of bad blocks
we should only skip over the area covered by the bad block.
RAID10 already gets this right, raid1 doesn't.  Or didn't.

As this can cause a 'sync' to abort early and appear to have succeeded
it could lead to some data corruption, so it suitable for -stable.

Cc: stable@vger.kernel.org
Reported-by: Alexander Lyakas <alex.bolshoy@gmail.com>
Signed-off-by: NeilBrown <neilb@suse.de>
(cherry picked from commit b7219ccb33aa0df9949a60c68b5e9f712615e56f)

Signed-off-by: Jes Sorensen <Jes.Sorensen@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>