hpsa: fix handling of protocol error
Patchwork-id: 51234
O-Subject: [RHEL6.4 PATCH] hpsa: fix handling of protocol error
Bugzilla: 857956
RH-Acked-by: Rob Evers <revers@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: David Milburn <dmilburn@redhat.com>

bz#857956

brew build - taskID=4880497

The patch is posted https://lkml.org/lkml/2012/9/14/557 ,
a similar patch for ccis was applied
by Jens Axboe - https://lkml.org/lkml/2012/9/18/164

Tomas

If a command status of CMD_PROTOCOL_ERR is received, this
information should be conveyed to the SCSI mid layer, not
dropped on the floor.  CMD_PROTOCOL_ERR may be received
from the Smart Array for any commands destined for an external
RAID controller such as a P2000, or commands destined for tape
drives or CD/DVD-ROM drives, if for instance a cable is
disconnected.  This mostly affects multipath configurations, as
disconnecting a cable on a non-multipath configuration is not
going to do anything good regardless of whether CMD_PROTOCOL_ERR
is handled correctly or not.  Not handling CMD_PROTOCOL_ERR
correctly in a multipath configaration involving external RAID
controllers may cause data corruption, so this is quite a serious
bug.  This bug should not normally cause a problem for direct
attached disk storage.

Signed-off-by: Stephen M. Cameron <scameron@beardog.cce.hp.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>