nfs: Prevent a deadlock in the new writeback code
Patchwork-id: 51247
O-Subject: [RHEL6.4 PATCH 08/12] NFS: Prevent a deadlock in the new writeback code
Bugzilla: 858357
RH-Acked-by: Jeff Layton <jlayton@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

From: Trond Myklebust <Trond.Myklebust@netapp.com>

We have to unlock the nfs_page before we call nfs_end_page_writeback
to avoid races with functions that expect the page to be unlocked
when PG_locked and PG_writeback are not set.
The problem is that nfs_unlock_request also releases the nfs_page,
causing a deadlock if the release of the nfs_open_context
triggers an iput() while the PG_writeback flag is still set...

The solution is to separate the unlocking and release of the nfs_page,
so that we can do the former before nfs_end_page_writeback and the
latter after.

Signed-off-by: Trond Myklebust <Trond.Myklebust@netapp.com>
Cc: Fred Isaman <iisaman@netapp.com>
Upstream Id: 3aff4ebb95b20ad8db2c1447e8c52097d89af5a7

Signed-off-by: Jarod Wilson <jarod@redhat.com>