thp: fix pmd_present for split_huge_page and PROT_NONE with THP
Patchwork-id: 51332
O-Subject: [RHEL6.4 PATCH] mm: thp: fix pmd_present for split_huge_page and PROT_NONE with THP
Bugzilla: 858385
RH-Acked-by: Johannes Weiner <jweiner@redhat.com>

In many places !pmd_present has been converted to pmd_none. For pmds
that's equivalent and pmd_none is quicker so using pmd_none is
better.

However (unless we delete pmd_present) we should provide an accurate
pmd_present too. This will avoid the risk of code thinking the pmd is
non present because it's under __split_huge_page_map, see the
pmd_mknotpresent there and the comment above it.

If the page has been mprotected as PROT_NONE, it would also lead to a
pmd_present false negative in the same way as the race with
split_huge_page.

Because the PSE bit stays on at all times (both during split_huge_page
and when the _PAGE_PROTNONE bit get set), we could only check for the
PSE bit, but checking the PROTNONE bit too is still good to remember
pmd_present must always keep PROT_NONE into account.

This explains a not reproducible BUG_ON that was seldom reported on
the lists.

The same issue is in pmd_large, it would go wrong with both PROT_NONE
and if it races with split_huge_page.

Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>