Hold a file reference in madvise_remove
Patchwork-id: 51344
O-Subject: [RHEL6.4 PATCH] CVE-2012-3511: mm: Hold a file reference in madvise_remove
Bugzilla: 849739
CVE: CVE-2012-3511
RH-Acked-by: Oleg Nesterov <oleg@redhat.com>
RH-Acked-by: Anton Arapov <Anton@redhat.com>

Bugzilla:
https://bugzilla.redhat.com/show_bug.cgi?id=849739

A race between madvise_remove() and munmap() or close() could allow an
unprivileged user to DOS a machine. The fix is to hold a file reference
in madvise_remove().
I wasn't able to reproduce the race and I'm not aware of any exploit in
the wild, but still the race seems plausible to me.

Backported from:
commit 4dd8e6ea1ef42aa9e3dcadc19dc0fc72d7c15df3
Author: Jerome Marchand <jmarchan@redhat.com>
Date:   Wed Sep 5 17:45:37 2012 +0200

    mm: Hold a file reference in madvise_remove

    Otherwise the code races with munmap (causing a use-after-free
    of the vma) or with close (causing a use-after-free of the struct
    file).

    The bug was introduced by commit 90ed52ebe481 ("[PATCH] holepunch: fix
    mmap_sem i_mutex deadlock")

    Cc: Hugh Dickins <hugh@veritas.com>
    Cc: Miklos Szeredi <mszeredi@suse.cz>
    Cc: Badari Pulavarty <pbadari@us.ibm.com>
    Cc: Nick Piggin <npiggin@suse.de>
    Cc: stable@vger.kernel.org
    Signed-off-by: Andy Lutomirski <luto@amacapital.net>
    Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>

Test status:
I've done some stress test with a lot of concurrent calls to madvise_remove(),
munmap() and close() (that is what I expected to be the reproducer).

Regards,
Jerome