Revert: Fix a merge issue
Patchwork-id: 51356
O-Subject: [Patch RHEL6 1/4] Revert "[infiniband] Fix a merge issue"
Bugzilla: 854053
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>

This reverts commit d15331b1fa34417038c0901cf17201c9b481906d.

Originally we had skipped the new steering code, so I had to create
a merge patch that compensated for that (6.2 timeframe).  Then we
integrated the new steering code (6.3 timeframe).  When I did that,
I didn't think about this.  Unfortunately, this merge was now
incorrect.  So, revert this merge as now that we have the new steering
code we need the code that this merge patch removed back.

This resolves bz854053.

Signed-off-by: Doug Ledford <dledford@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>