virtuoso: fix front panel routing for D1/DX/ST(X)
Patchwork-id: 51565
O-Subject: [RHEL 6.4 PATCH 36/83] ALSA/CMI878x: virtuoso: fix front panel routing for D1/DX/ST(X)
Bugzilla: 816880
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

The "Front Panel" switch on the Xonar D1/DX actually switches only the
output direction, so mark it appropriately.

The front panel microphone is controlled by the FMIC2MIC bit of the
CM9780.  It was unconditionally enabled on the D1/DX and never set on
the ST(X); add a control for it.  Selecting the front panel microphone
as source does not actually disable the microphone jack, but this is
bug-compatible with the Windows driver, and users rely on it.

Signed-off-by: Clemens Ladisch <clemens@ladisch.de>
Signed-off-by: Takashi Iwai <tiwai@suse.de>
Signed-off-by: Jaroslav Kysela <perex@perex.cz>

(picked from kernel: e96f38f732d24515792296b3738842934c985539)
rhbz#816880

Signed-off-by: Jarod Wilson <jarod@redhat.com>