virtuoso: fix setting of Xonar DS line-in/mic-in controls
Patchwork-id: 51560
O-Subject: [RHEL 6.4 PATCH 38/83] ALSA/CMI878x: virtuoso: fix setting of Xonar DS line-in/mic-in controls
Bugzilla: 816880
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

The Line and Mic inputs cannot be used at the same time, so the driver
has to automatically disable one of them if both are set.  However, it
forgot to notify userspace about this change, so the mixer state would
be inconsistent.  To fix this, check if the other control gets muted,
and send a notification event in this case.

Signed-off-by: Clemens Ladisch <clemens@ladisch.de>
Reported-and-tested-by: Nathan Schagen
Cc: <stable@kernel.org>
Signed-off-by: Takashi Iwai <tiwai@suse.de>
Signed-off-by: Jaroslav Kysela <perex@perex.cz>

(picked from kernel: fe6ce80ae25953d95ebaf9bce27b585218cda25c)
rhbz#816880

Signed-off-by: Jarod Wilson <jarod@redhat.com>