hda: Revert "ALSA: hda - Set codec to D3 forcibly even if not used"
Patchwork-id: 51697
O-Subject: [RHEL 6.4 PATCH 100/245] ALSA/HDA: Revert "ALSA: hda - Set codec to D3 forcibly even if not used"
Bugzilla: 838558
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

ALSA/HDA: Revert "ALSA: hda - Set codec to D3 forcibly even if not used"

Message-id: <201209211219.q8LCJ54o026633@int-mx09.intmail.prod.int.phx2.redhat.com>
Patchwork-id: 51697
O-Subject: [RHEL 6.4 PATCH 100/245] ALSA/HDA: Revert "ALSA: hda - Set codec to D3 forcibly even if not used"
Bugzilla: 838558

This reverts commit 785f857d1cb0856b612b46a0545b74aa2596e44a.

The commit causes a problem with the wrong D3 state after suspend
because the call of hda_set_power_state() involves with the power-up
sequence, which changes the power_count, and this confuses the resume
sequence that checks the power_count as well.

Originally, this go-to-D3 sequence should be a simple task without the
power-up sequence.  But, it'd need some proper sanity checks in the
case of power-saved state, so it's not too easy to write now in the
3.4-rc cycle.

In short, the safest option now is to revert this affecting commit.

Of course, we need to clean up and robustify the power-saving code
better for 3.5 kernel.

Reported-by: Konstantin Khlebnikov <khlebnikov@openvz.org>
Signed-off-by: Takashi Iwai <tiwai@suse.de>

(picked from kernel: 619a341b78f17fb86d92e89c04612676cd05e26f)
rhbz#838558