usbaudio: Support for Power/Status LED on Creative USB X-Fi S51
Patchwork-id: 51995
O-Subject: [RHEL 6.4 PATCH 112/268] ALSA/USB: usb-audio - Support for Power/Status LED on Creative USB X-Fi S51
Bugzilla: 664586
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

[alsa] usb-audio - Support for Power/Status LED on Creative USB X-Fi S51

Message-id: <201209211227.q8LCRm2o022870@int-mx02.intmail.prod.int.phx2.redhat.com>
Patchwork-id: 51995
O-Subject: [RHEL 6.4 PATCH 112/268] ALSA/USB: usb-audio - Support for Power/Status LED on Creative USB X-Fi S51
Bugzilla: 664586

This patch adds support for Power/Status LED on Creative USB X-Fi S51.
There is just one LED on the device. The LED can either be On or it
can be set to Blink. There doesn't seem to be a way to switch it off.
The control message to change LED status is similar to that of
audigy2nx except that the index is to be set to 0 and value is 1 for
Blink and 0 for On.

The 'Power LED' control in alsamixer when muted will cause the LED to
Blink continuously. When unmuted  the LED will stay On. The Creative
driver under Windows sets the LED to blink whenever audio is muted.
This LED can be treated as the CMSS LED but I figured since there is
just one LED, it should be treated as the Power LED. Is that alright?

I've also changed the comment "Usb X-Fi" to "Usb X-Fi S51" as there
are other external X-Fi devices from Creative like Usb X-Fi Go and
Xmod. The volume knob and LED support patch doesn't apply to them.

Signed-off-by: Mandar Joshi <emailmandar@gmail.com>
Signed-off-by: Takashi Iwai <tiwai@suse.de>

(picked from kernel: ca8dc34eaf7a1db7daa604495ed2c143af32f1ed)
rhbz#664586