bnx2x: Look inside vlan when determining checksum proto.
Patchwork-id: 52130
O-Subject: [RHEL6.4 PATCH 002/141] bnx2x: Look inside vlan when determining checksum proto.
Bugzilla: 819842 826277 826295 826302
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>

From: Hao Zheng <hzheng@nicira.com>

Currently the skb->protocol field is used to setup checksum
offloading on transmit for the correct protocol.  However, if
vlan offloading is disabled or otherwise not used, the protocol
field will be ETH_P_8021Q, not the actual protocol.  This will
cause the checksum to be not computed correctly, even though the
hardware is capable of looking inside vlan tags.  Instead,
look inside the header if necessary to determine the correct
protocol type.

To some extent this fixes a regression from 2.6.36 because it
was previously not possible to disable vlan offloading and this
error case was not exposed.

Signed-off-by: Hao Zheng <hzheng@nicira.com>
CC: Eilon Greenstein <eilong@broadcom.com>
Signed-off-by: Jesse Gross <jesse@nicira.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit d0d9d8ef5949551276f635cb04969184ba1a9553)

Signed-off-by: Jarod Wilson <jarod@redhat.com>