ext4: fix start and len arguments handling in ext4_trim_fs()
Patchwork-id: 52335
O-Subject: [RHEL6.4 PATCH 1/3] ext4: fix start and len arguments handling in ext4_trim_fs()
Bugzilla: 858012
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>
RH-Acked-by: Carlos Maiolino <cmaiolino@redhat.com>

BZ 858012
https://bugzilla.redhat.com/show_bug.cgi?id=858012

BREW 4894633
https://brewweb.devel.redhat.com/taskinfo?taskID=4894633

This is backport of upstream commit
913eed83eda42012cde8a9a801c31b312545c098. Due to bigalloc changes some
changes were required. Most of the work has been done in rhel commit
7179b80ce33ff20faf5c0a1ccac7962490213d19 already so we just backport
missineg pieces.

The overflow can happen when we are calling get_group_no_and_offset()
which stores the group number in the ext4_grpblk_t type which is
actually int. However when the blocknr is big enough the group number
might be bigger than ext4_grpblk_t resulting in overflow. This will
most likely happen with FITRIM default argument len = ULLONG_MAX.

Fix this by using "end" variable instead of "start+len" as it is easier
to get right and specifically check that the end is not beyond the end
of the file system, so we are sure that the result of
get_group_no_and_offset() will not overflow. Otherwise truncate it to
the size of the file system.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>
Reviewed-by: Jan Kara <jack@suse.cz>
Signed-off-by: "Theodore Ts'o" <tytso@mit.edu>
Signed-off-by: Jarod Wilson <jarod@redhat.com>