gfs2: Remove rs_requested field from reservations
Patchwork-id: 52481
O-Subject: [RHEL6.4 PATCH 01/23] GFS2: Remove rs_requested field from reservations
Bugzilla: 848109
RH-Acked-by: Abhijith Das <adas@redhat.com>
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>

From: Bob Peterson <rpeterso@redhat.com>

The rs_requested field is left over from the original allocation
code, however this should have been a parameter passed to the
various functions from gfs2_inplace_reserve() and not a member of the
reservation structure as the value is not required after the
initial allocation.

This also helps simplify the code since we no longer need to set
the rs_requested to zero. Also the gfs2_inplace_release()
function can also be simplified since the reservation structure
will always be defined when it is called, and the only remaining
task is to unlock the rgrp if required. It can also now be
called unconditionally too, resulting in a further simplification.

rhbz#848109

Signed-off-by: Jarod Wilson <jarod@redhat.com>