gfs2: Add structure to contain rgrp, bitmap, offset tuple
Patchwork-id: 52497
O-Subject: [RHEL6.4 PATCH 14/23] GFS2: Add structure to contain rgrp, bitmap, offset tuple
Bugzilla: 848109
RH-Acked-by: Abhijith Das <adas@redhat.com>
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>

From: Bob Peterson <rpeterso@redhat.com>

In addition, the rbtree nodes which are used for the reservations
were not being correctly initialised, which is now fixed. Also,
the tracing was not passing through the inode where it should
have been. That is mostly fixed aside from one corner case. This
needs to be revisited since there can also be a NULL rgrp in
some cases which results in the device being incorrect in the
trace.

This is intended to be the first step towards cleaning up some
of the allocation code, and some further bug fixes.

rhbz#848109

Signed-off-by: Jarod Wilson <jarod@redhat.com>