gfs2: Use rbm for gfs2_setbit()
Patchwork-id: 52496
O-Subject: [RHEL6.4 PATCH 19/23] GFS2: Use rbm for gfs2_setbit()
Bugzilla: 848109
RH-Acked-by: Abhijith Das <adas@redhat.com>
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>

From: Bob Peterson <rpeterso@redhat.com>

Use the rbm structure for gfs2_setbit() in order to simplify the
arguments to the function. We have to add a bool to control whether
the clone bitmap should be updated (if it exists) but otherwise it
is a more or less direct substitution.

rhbz#848109

Signed-off-by: Jarod Wilson <jarod@redhat.com>