gfs2: Fall back to ignoring reservations, if there are no other blocks left
Patchwork-id: 52502
O-Subject: [RHEL6.4 PATCH 20/23] GFS2: Fall back to ignoring reservations, if there are no other blocks left
Bugzilla: 848109
RH-Acked-by: Abhijith Das <adas@redhat.com>
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>

From: Bob Peterson <rpeterso@redhat.com>

When we get to the stage of allocating blocks, we know that the
resource group in question must contain enough free blocks, otherwise
gfs2_inplace_reserve() would have failed. So if we are left with only
free blocks which are reserved, then we must use those. This can happen
if another node has sneaked in and use some blocks reserved on this
node, for example. Generally this will happen very rarely and only
when the resource group is nearly full.

rhbz#848109

Signed-off-by: Jarod Wilson <jarod@redhat.com>