e1000e: rename e1000e_config_collision_dist() and call as function pointer
Patchwork-id: 52538
O-Subject: [RHEL6.4 PATCH 31/84] e1000e: rename e1000e_config_collision_dist() and call as function pointer
Bugzilla: 819103
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jiri Pirko <jpirko@redhat.com>

Resolves BZ:
https://bugzilla.redhat.com/show_bug.cgi?id=819103

Upstream status:
This is a backport of the following from the linux tree...

commit 57cde7630c1911ea7e8e1561cccfde8096e8bcc7
Author: Bruce Allan <bruce.w.allan@intel.com>
Date:   Wed Feb 22 09:02:58 2012 +0000

    e1000e: rename e1000e_config_collision_dist() and call as function pointer

    Rename e1000e_config_collision_dist() to
    e1000e_config_collision_dist_generic() to signify the function is used for
    more than one MAC-family type, and set and use it as a MAC ops function
    pointer to be consistent with the driver design.

    Signed-off-by: Bruce Allan <bruce.w.allan@intel.com>
    Tested-by: Aaron Brown <aaron.f.brown@intel.com>
    Signed-off-by: Jeff Kirsher <jeffrey.t.kirsher@intel.com>

The difference between the following patch and the upstream commit is due
to filename differences, upstream's 80003es2lan.c versus RHEL6's es2lan.c.
And to not having backported dbf80dcbd8ca0c50f343 which is dependent
on .set_phys_id, something that is not defined in RHEL6.

Signed-off-by: Jarod Wilson <jarod@redhat.com>