e1000e: remove use of IP payload checksum
Patchwork-id: 52577
O-Subject: [RHEL6.4 PATCH 70/84] e1000e: remove use of IP payload checksum
Bugzilla: 819103
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jiri Pirko <jpirko@redhat.com>

Resolves BZ:
https://bugzilla.redhat.com/show_bug.cgi?id=819103

Upstream status:
This is a partial backport of the following from the linux tree...

commit 2e1706f234f86ff71056ef69683d734fbf7e9e40
Author: Bruce Allan <bruce.w.allan@intel.com>
Date:   Sat Jun 30 20:02:42 2012 +0000

    e1000e: remove use of IP payload checksum

    Currently only used when packet split mode is enabled with jumbo frames,
    IP payload checksum (for fragmented UDP packets) is mutually exclusive with
    receive hashing offload since the hardware uses the same space in the
    receive descriptor for the hardware-provided packet checksum and the RSS
    hash, respectively.  Users currently must disable jumbos when receive
    hashing offload is enabled, or vice versa, because of this incompatibility.
    Since testing has shown that IP payload checksum does not provide any real
    benefit, just remove it so that there is no longer a choice between jumbos
    or receive hashing offload but not both as done in other Intel GbE drivers
    (e.g. e1000, igb).

    Also, add a missing check for IP checksum error reported by the hardware;
    let the stack verify the checksum when this happens.

    CC: stable <stable@vger.kernel.org> [3.4]
    Signed-off-by: Bruce Allan <bruce.w.allan@intel.com>
    Tested-by: Aaron Brown <aaron.f.brown@intel.com>
    Signed-off-by: Jeff Kirsher <jeffrey.t.kirsher@intel.com>
    Signed-off-by: David S. Miller <davem@davemloft.net>

The difference between the following patch and the upstream commit is due
to RHEL6 not having the receive hashing offload feature, thus making the
last two hunks of the upstream commit irrelevant. As well, as it not being
a good idea to remove the IP payload checksumming feature, which eliminates
all the other hunks except two. These two add a missing check for IP
checksum error mentioned in the upstream's changelog. A blank line added by
upstream's afd12939a09ca8f96cf8 (which wasn't backported to RHEL6) affects
the context of the second hunk of this patch.

Signed-off-by: Jarod Wilson <jarod@redhat.com>