xfs: fix possible overflow in xfs_ioc_trim()
Patchwork-id: 52661
O-Subject: [RHEL6.4 PATCH 011/109] xfs: fix possible overflow in xfs_ioc_trim()
Bugzilla: 700324
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>

From: Lukas Czerner <lczerner@redhat.com>

Upstream commit: c029a50d51b8a9520105ec903639de03389915d0
RH BZ: 753224
In xfs_ioc_trim it is possible that computing the last allocation group
to discard might overflow for big start & len values, because the result
might be bigger then xfs_agnumber_t which is 32 bit long. Fix this by not
allowing the start and end block of the range to be beyond the end of the
file system.

Note that if the start is beyond the end of the file system we have to
return -EINVAL, but in the "end" case we have to truncate it to the fs
size.

Also introduce "end" variable, rather than using start+len which which
might be more confusing to get right as this bug shows.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>
Reviewed-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: Alex Elder <aelder@sgi.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>