xfs: remove the i_size field in struct xfs_inode
Patchwork-id: 52677
O-Subject: [RHEL6.4 PATCH 041/109] xfs: remove the i_size field in struct xfs_inode
Bugzilla: 845233
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>

From: Christoph Hellwig <hch@infradead.org>

Upstream commit: ce7ae151ddada3dbf67301464343c154903166b3
RH BZ: 845233
There is no fundamental need to keep an in-memory inode size copy in the XFS
inode.  We already have the on-disk value in the dinode, and the separate
in-memory copy that we need for regular files only in the XFS inode.

Remove the xfs_inode i_size field and change the XFS_ISIZE macro to use the
VFS inode i_size field for regular files.  Switch code that was directly
accessing the i_size field in the xfs_inode to XFS_ISIZE, or in cases where
we are limited to regular files direct access of the VFS inode i_size field.

This also allows dropping some fairly complicated code in the write path
which dealt with keeping the xfs_inode i_size uptodate with the VFS i_size
that is getting updated inside ->write_end.

Note that we do not bother resetting the VFS i_size when truncating a file
that gets freed to zero as there is no point in doing so because the VFS inode
is no longer in use at this point.  Just relax the assert in xfs_ifree to
only check the on-disk size instead.

Reviewed-by: Dave Chinner <dchinner@redhat.com>
Signed-off-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: Ben Myers <bpm@sgi.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>