xfs: fallback to vmalloc for large buffers in xfs_getbmap
Patchwork-id: 52699
O-Subject: [RHEL6.4 PATCH 069/109] xfs: fallback to vmalloc for large buffers in xfs_getbmap
Bugzilla: 858353
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>

Upstream commit: f074211f6041305b645669464343d504f4e6a290
RH BZ: 858353
xfs_getbmap uses for a large buffer for extents, which is kmalloc'd.
This can fail after the system has been running for some time as it
is a high order allocation. Add a fallback to vmalloc so that it
doesn't require contiguous memory and so won't randomly fail on
files with large extent lists.

Signed-off-by: Dave Chinner <dchinner@redhat.com>
Reviewed-by: Christoph Hellwig <hch@lst.de>
Reviewed-by: Mark Tinguely <tinguely@sgi.com>
Signed-off-by: Ben Myers <bpm@sgi.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>